Add subProp case to Seq aliases suite
